requirements.txt
numpy>=1.26.0
google-cloud-storage

